Finished benchmarking, began BVH
Finished benchmarking. Turns out I implemented the
search from PS_DTFE poorly, so I was just measuring
the time for julia to throw an error, which is
why it was orders of magnitude better.

Starting work on BVH. I will refactor it more
as opposed to the PS_DTFE implementation.